No se genera ningún archivo hasta que no se seleccione una carpeta de destino

+++ app/debitos/generar-debitos.component.js
@@ -15,26 +15,27 @@ angular.
 
             document.getElementById("upload").addEventListener('change', function () {
               var dir = this.value;
-              FileAPI.copyDir('app/resources/tmp/', dir)
-                .then(function () {
-                    FileAPI.deleteDir('app/resources/tmp');
-                    window.alert("Los archivos se han generado exitosamente!");
-                })
-                .catch(function (error) {
-                  console.log(error);
-                });
+              if (dir) {
+                debitosService.exportarArchivo(self.entidad)
+                  .then(function() {
+                      FileAPI.copyDir('app/resources/tmp/', dir)
+                        .then(function () {
+                            FileAPI.deleteDir('app/resources/tmp');
+                            window.alert("Los archivos se han generado exitosamente!");
+                        })
+                        .catch(function (error) {
+                          console.log(error);
+                        });
+                  })
+                  .catch(function(error) {
+                        alert(error);
+                  });
+              }
             });
           };
 
           this.generar = function() {
-            debitosService.exportarArchivo(self.entidad)
-              .then(function() {
-                      document.getElementById('upload').click();
-                  })
-              .catch(function(error) {
-                    alert(error);
-                 });
-
+            document.getElementById('upload').click();
          };
 
       }]
